training/autoencoder/requirements.txt
tensorflow<2.11
matplotlib==3.7.1
pandas==2.0.1
Pillow==9.5.0
scikit-image==0.21.0
radiant-mlhub==0.5.5
rasterio==1.3.7
scikit-learn==1.3.0
scikit-image==0.21.0
